Modify PMIC power button to not suspend the interrupt during suspend/resume
The PMIC power button interrupt routine does something like:

1) Interrupt occurs (Falling or rising edge)
2) I2C transaction to PMIC to find current button level
3) Issue press/release event based on line-level of button

The timing between steps 1 and 2 are extremely sensitive.  This needs to occur
rather quickly to get accurate results.  In order to mitigate this when the
power button is pressed during wakeup, make sure that the driver can
immediately query the PMIC even during suspend/resume.  The driver can't
get an interrupt unless the I2C driver is enabled, so this should be mostly
safe.  There might be a small race-condition window while suspending
and with the button pressed at this point however.

DM3730ANDROID-117

git-svn-id: svn+ssh://svn/cvs/eps_svn/software/products/linux/kernels/linux-3.0@28763 d7e73fac-2431-0410-abd5-fd325483733a