Changes to allow EHCI controller to enter off (or retention) mode
Ideally, we'd be able to use the SAR (Save and restore) mechanism
for resuming the USB controller.  Unfortunately, this is fraught
with glitches if it is used (sometimes bad values are restored
and the entire EHCI controller simply has issues at that point).
To note, SAR is only applicable to going into OFF-mode, and not
into retention mode.

If you'd like to experiment with the SAR restore mechanism,
modify the structure "usbhost_pwrdm" within the file
"arch/arm/mach-omap2/powerdomains3xxx_data.c" to have the
flag "PWRDM_HAS_HDWR_SAR" be set (you'll see it commented
out in the source file).

Otherwise, the workaround for the off-mode glitch is to currently
fully reset the EHCI controller at every resume where the context
was detected to be lost (retention mode doesn't need this
workaround).  This will cause all USB devices to disconnect
and reconnect, including the built-in hub on the LV_SOM.
Additionally, one _could_ indicate to the EHCI controller that
the root hub power has been lost, however, this unfortunatley
adds over a second into the resume path before userspace is running.

DM37LINUX-592

git-svn-id: svn+ssh://svn/cvs/eps_svn/software/products/linux/kernels/linux-3.0@28900 d7e73fac-2431-0410-abd5-fd325483733a