USB: TWL4030: Add polling while ID pin is low
The USB_PRES interrupt that the TWL4030 relies on for detection of
ID pin and VBUS changes, is actually a combination of two different
logical signals.  USB_PRES = (ID_PRES | VBUS_PRES).  The implication
of this is that when a session is active in the OTG controller,
it actually raises triggers the VBUS_PRES portion of the interrupt.

Sequence of events for what this fixes:

1. Plug in cable with ID pin grounded
2. ID_PRES goes high
3. USB_PRES interrupt is triggered due to ID_PRES
4. OTG controller enables VBUS
5. VBUS_PRES goes high
6. User unplugs cable
7. ID pin no longer grounded, ID_PRES goes low, however,
   USB_PRES is still true because of step 4 where the
   controller enabled VBUS.

Essentially, step 7 _should_ change the type of connection due
to the ID pin going low, however, it can't do this with just
interrupt conditions because of the combination of signals that
create the interrupt.  We're reduced to doing polling in order to
properly create an event for the ID pin floating at this point.
This patch only polls while the ID pin was previously true.  It also
modifies the behavior of ID_EVENT_* to trigger ID_EVENT_VBUS only
when the OTG controller hadn't specifically turned on VBUS as well.

Signed-off-by: Tim Nordell <tim.nordell@logicpd.com>

git-svn-id: svn+ssh://svn/cvs/eps_svn/software/products/linux/kernels/linux-3.0@28979 d7e73fac-2431-0410-abd5-fd325483733a