ISP1763: Remote wakeup fixed for root hub port
Various issues in the implementation of the suspend/resume of the
root hub port on the ISP1763 caused wakeups to not be seen by the
USB hub layer.  Fixed these, and now the USB ports can wakeup
properly.

(To be noted is that as a side-effect is that the entire system
now wakes up when a USB device is inserted or removed.  This
can be disabled by changing the wakeup field for the USB host
device in power [e.g. something like
/sys/bus/usb/devices/usb1/power/wakeup] to disabled instead of
the default of enabled for Linux.  Additionally, things like
pressing a key on a keyboard attached to USB will wakeup the
system as well unless the equivalent wakeup is disabled.)

DM37LINUX-517
DM3730ANDROID-281

git-svn-id: svn+ssh://svn/cvs/eps_svn/software/products/linux/kernels/linux-3.0@29016 d7e73fac-2431-0410-abd5-fd325483733a