Shared Transport: Enable/disable uart timeout based on shared transport activity.
Note: This is sort of a hackish hack - it shouldn't directly assume the UART
that is being used (e.g. it assumes the shared transport is attached to
ttyO1 exactly - if it isn't, this will end up enabling/disabling the wrong
port's timeout).  Fortunately, we have exact control over the UART being used,
and we're assuming this patch will never make it upstream.  If it is to make
it upstream, the mechanism for this patch would be quite different (upstream
has PM runtime support) as there is better power management for the UART
driver in upstream.